refactor: 레시피 폼 제출 로직을 별도 함수로 분리
- 폼 제출 로직의 복잡성을 낮추기 위해 submitRecipeForm 함수로 분리
- 이미지 업로드 로직을 포함한 폼 제출 처리를 모듈화하여 재사용성 향상
